chore: moved the scroll to a hook
